[FIX] Do not import delaunator in tests
Splits asset graphics calculation functions into a separate file to achieve that
