+++ pandora-client-web/src/editor/components/assets/assets.tsx
@@ -2,25 +2,26 @@ import classNames from 'classnames';
 import { Assert, AssertNotNullable, Asset, AssetId, Item } from 'pandora-common';
 import React, { ReactElement, useCallback, useState, useSyncExternalStore } from 'react';
 import { useForm, Validate } from 'react-hook-form';
-import { AssetGraphicsLayer, useLayerName } from '../../../assets/assetGraphics';
+import { toast } from 'react-toastify';
+import { AssetGraphicsLayer } from '../../../assets/assetGraphics';
+import { useLayerName } from '../../../assets/assetGraphicsCalculations';
 import { useCharacterAppearanceItems } from '../../../character/character';
 import { Button } from '../../../components/common/button/button';
 import { Row } from '../../../components/common/container/container';
 import { Form, FormField, FormFieldError } from '../../../components/common/form/form';
-import { Select } from '../../../components/common/select/select';
 import { Scrollbar } from '../../../components/common/scrollbar/scrollbar';
+import { Select } from '../../../components/common/select/select';
 import { ModalDialog } from '../../../components/dialog/dialog';
 import { ContextHelpButton } from '../../../components/help/contextHelpButton';
 import { StripAssetIdPrefix } from '../../../graphics/utility';
-import { IObservableClass, ObservableProperty, ObservableClass, useObservableProperty } from '../../../observable';
-import { AssetManagerEditor, AssetTreeViewCategory, ASSET_ID_PART_REGEX, useAssetManagerEditor } from '../../assets/assetManager';
+import { IObservableClass, ObservableClass, ObservableProperty, useObservableProperty } from '../../../observable';
+import { TOAST_OPTIONS_ERROR } from '../../../persistentToast';
+import { ASSET_ID_PART_REGEX, AssetManagerEditor, AssetTreeViewCategory, useAssetManagerEditor } from '../../assets/assetManager';
 import { EDITOR_ALPHA_ICONS, useEditorLayerTint, useEditorTabContext } from '../../editor';
 import { useEditor } from '../../editorContextProvider';
-import './assets.scss';
-import { toast } from 'react-toastify';
-import { TOAST_OPTIONS_ERROR } from '../../../persistentToast';
 import { useEditorCharacterState } from '../../graphics/character/appearanceEditor';
 import { PreviewCutter } from '../previewCutter/previewCutter';
+import './assets.scss';
 
 export function AssetsUI(): ReactElement {
 	const editor = useEditor();
