+++ pandora-client-web/src/editor/editor.tsx
@@ -3,7 +3,8 @@ import { noop } from 'lodash';
 import { Assert, AssertNotNullable, AssetFrameworkCharacterState, AssetFrameworkGlobalState, AssetFrameworkGlobalStateContainer, AssetFrameworkRoomState, AssetId, CharacterSize, GetLogger, HexColorString, ParseArrayNotEmpty, TypedEventEmitter, type PointTemplate } from 'pandora-common';
 import React, { createContext, ReactElement, useContext, useMemo, useSyncExternalStore } from 'react';
 import z from 'zod';
-import { AssetGraphics, AssetGraphicsLayer, useGraphicsAsset, useLayerDefinition } from '../assets/assetGraphics';
+import { AssetGraphics, AssetGraphicsLayer } from '../assets/assetGraphics';
+import { useGraphicsAsset, useLayerDefinition } from '../assets/assetGraphicsCalculations';
 import { GetCurrentAssetManager } from '../assets/assetManager';
 import { GraphicsManager } from '../assets/graphicsManager';
 import { useBrowserStorage } from '../browserStorage';
