+++ pandora-client-web/src/editor/graphics/layer/setupLayer.tsx
@@ -3,7 +3,8 @@ import { AssetFrameworkCharacterState } from 'pandora-common';
 import * as PIXI from 'pixi.js';
 import { Texture } from 'pixi.js';
 import React, { ReactElement, useCallback, useEffect, useMemo, useReducer } from 'react';
-import { AssetGraphicsLayer, useLayerDefinition, useLayerImageSource } from '../../../assets/assetGraphics';
+import { AssetGraphicsLayer } from '../../../assets/assetGraphics';
+import { useLayerDefinition, useLayerImageSource } from '../../../assets/assetGraphicsCalculations';
 import { useCharacterAppearanceItems } from '../../../character/character';
 import { useAppearanceConditionEvaluator } from '../../../graphics/appearanceConditionEvaluator';
 import { GraphicsLayerProps, useItemColor, useLayerPoints, useLayerVertices } from '../../../graphics/graphicsLayer';
